jgourd: tweaked reverse engineering build scripts and better handled AMD64/ARM64 versions of lab07.

+++ labs/09 Reverse Engineering/examples/clean.sh
@@ -5,5 +5,8 @@ for f in *.c; do
     fn=$(basename -- "$f")
     fe="${fn##*.}"
     fn="${fn%.*}"
-    rm $fn
+
+    if [ -f "$fn" ]; then
+        rm $fn
+    fi
 done
